fixed all the warnings on Cray with the "ftn" compiler (except warnings about temporary array copies). In particular, renamed sync_all() to synchronize_all() because sync_all() seems to be a reserved function on Cray.

+++ src/meshfem3D/model_gapp2.f90
@@ -215,11 +215,11 @@
   if(theta>=PI) then
      ia = na
   else
-     ia = theta / dtheta + 1
+     ia = int(theta / dtheta) + 1
   endif
   ! longitude
   if(phi < 0.0d0) phi = phi + 2.*PI
-  io=phi / dphi + 1
+  io=int(phi / dphi) + 1
   if(io>no) io=io-no
 
   ! velocity and density perturbations
